chore: delete dormant AI commands, fold openai_integration into ai_providers; docs; bump 2.5.0
- Removes the dormant pre-2.3 AI surface (analyze_with_openai,
  analyze_system_with_ai with its forced-5-10-diagnostics prompt,
  analyze_system_with_phi_silica, analyze_system_with_ai_provider,
  get_ai_provider_status, parse_analysis and friends) — zero frontend
  references; the agentic chat replaces all of it.
- The survivors (OPENAI_MODEL, FOUNDRY_LOCAL_MODEL, Foundry endpoint
  resolution) move into ai_providers/{openai,foundry}.rs and
  openai_integration.rs is gone. dpapi compiles cross-platform again
  (it self-gates internally and ProviderKeyId names keyring entries).
- CLAUDE.md: seven-provider capability table, routing rules, encoded
  provider gotchas (Anthropic max_tokens/refusal, Gemini header auth /
  functionResponse object, chat-completions vs responses split), key
  storage scheme, chat/report command + event reference, and the
  read-only tool rule.
- Store release notes for 2.5.0; version bumped across all 10 files.
