+++ AppxManifest.xml
@@ -10,7 +10,7 @@
   <Identity
     Name="32827MikeFara.WindowsForumDiagnostics"
     Publisher="CN=ABDB6B3F-DF9E-447D-BC0E-4DA7BAFD14C4"
-    Version="2.4.0.0"
+    Version="2.5.0.0"
     ProcessorArchitecture="x64" />
 
   <mp:PhoneIdentity PhoneProductId="32827e7c-77d2-43e5-ab82-9cdb9daa11b3" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
@@ -23,11 +23,11 @@
 
   <Dependencies>
     <!-- Both Universal and Desktop required for WinRT AI APIs (Phi Silica) -->
-    <TargetDeviceFamily Name="Windows.Universal" MinVersion="2.4.0.0" MaxVersionTested="10.0.26226.0" />
-    <TargetDeviceFamily Name="Windows.Desktop" MinVersion="2.4.0.0" MaxVersionTested="10.0.26226.0" />
+    <TargetDeviceFamily Name="Windows.Universal" MinVersion="2.5.0.0" MaxVersionTested="10.0.26226.0" />
+    <TargetDeviceFamily Name="Windows.Desktop" MinVersion="2.5.0.0" MaxVersionTested="10.0.26226.0" />
     <!-- Windows App SDK 2.0 experimental3 for Phi Silica AI APIs -->
     <PackageDependency Name="Microsoft.WindowsAppRuntime.2.0-experimental3"
-                       MinVersion="2.4.0.0"
+                       MinVersion="2.5.0.0"
                        Publisher="CN=Microsoft Corporation, O=Microsoft Corporation, L=Redmond, S=Washington, C=US" />
   </Dependencies>
