+++ package.json
@@ -1,7 +1,7 @@
 {
     "name": "wfdiag-tauri-ui",
     "private": true,
-    "version": "2.4.0",
+    "version": "2.5.0",
     "type": "module",
     "scripts": {
         "dev": "vite",
