+++ src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wfdiag-tauri"
-version = "2.4.0"
+version = "2.5.0"
 description = "WindowsForum Diagnostic Tool"
 authors = ["WindowsForum"]
 edition = "2024"
