+++ src-tauri/tauri.conf.json
@@ -22,7 +22,7 @@
                                }
                },
     "productName":  "WindowsForum Diagnostics",
-    "version": "2.4.0",
+    "version": "2.5.0",
     "identifier":  "com.windowsforum.diagnostics",
     "plugins":  {
