+++ version.json
@@ -1,5 +1,5 @@
 {
-    "version": "2.4.0",
+    "version": "2.5.0",
     "name": "WF Diagnostics",
     "description": "WindowsForum Diagnostic Tool"
 }
